Recreate MainViewModel and move several CleverTap SDK set up from FeaturesViewModel.kt to MainViewModel.kt.
