+++ app/src/main/java/com/heyzeusv/clevertapassessment/di/AppModule.kt
@@ -4,6 +4,8 @@ import com.clevertap.android.sdk.CleverTapAPI
 import com.clevertap.android.sdk.CleverTapAPI.LogLevel
 import com.clevertap.android.sdk.CleverTapAPI.setDebugLevel
 import com.heyzeusv.clevertapassessment.BuildConfig
+import com.heyzeusv.clevertapassessment.ui.MainViewModel
+import com.heyzeusv.clevertapassessment.ui.eventform.EventFormViewModel
 import com.heyzeusv.clevertapassessment.ui.features.FeaturesViewModel
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.module.dsl.viewModelOf
@@ -20,5 +22,7 @@ val appModule = module {
 			}
 		}
 	}
+	viewModelOf(::MainViewModel)
 	viewModelOf(::FeaturesViewModel)
+	viewModelOf(::EventFormViewModel)
 }
